1. Move all audio scripts into the MusicManager folder 2. rewrite the UIButtonSound script, so that after the UI interface is closed, it can still play the sound effect completely. 3. Changed WindZoomOut sound size and Building_Click_LoggingCamp sound. 4. Configured PeasAndBeans prefab sound effect. 5. Added Building_Click_Standingstone, Building_Click_SpringBarley, Building_Place_SpringBarley, sound effects. 6. Removed duplicate UIButtonSound component from prefab.
