+++ Assets/Audio/SFX/Building_Click_Standingstone.WAV
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:397ce7039a518b4f85b150ac8d957f10d679b066b883156f47ea9b7e65bd3acf
+size 458830
