+++ Assets/Audio/SFX/Building_Place_LoggingCamp.WAV
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:6edf1885031f39bd4136e290eed25a6fdbad33cbb7a8b44b560655f23d1539e4
-size 532558
+oid sha256:f8faed115d0ca12ee836ece4c4725fd79509540f223df60d062bfb3143a688fd
+size 561230
