+++ Assets/Prefabs/UI/StandingStoneUI.prefab
@@ -265,6 +265,7 @@ GameObject:
   - component: {fileID: 5268200023604658286}
   - component: {fileID: 1573395792918828394}
   - component: {fileID: 1942026777991167568}
+  - component: {fileID: 7238951490700715368}
   m_Layer: 5
   m_Name: Button
   m_TagString: Untagged
@@ -386,6 +387,21 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
+--- !u!114 &7238951490700715368
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4474113171628837302}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9d340082af1f8c947ab943b16d34a83f, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  ClickedSound: {fileID: 8300000, guid: d8c5aa096df004043aab188d53513cdb, type: 3}
+  mixerGroup: {fileID: 777363791655524067, guid: 431f31701a6ba7f438d79de92b1aeef7,
+    type: 2}
 --- !u!1 &4779920850894636769
 GameObject:
   m_ObjectHideFlags: 0
